requirements.txt
pandas
groq
google-genai
scikit-learn
python-dotenv
kagglehub[pandas-datasets]
vertexai